Admin etc

+++ HabitAqui/Areas/Identity/Pages/Account/Manage/ApagarCategoria.cshtml.cs
@@ -1,6 +1,7 @@
 using HabitAqui.Data;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.EntityFrameworkCore;
 
 namespace HabitAqui.Areas.Identity.Pages.Account.Manage
 {
@@ -10,7 +11,7 @@ namespace HabitAqui.Areas.Identity.Pages.Account.Manage
 
         [BindProperty]
         public int CategoriaId { get; set; }
-
+        public string CategoriaNome { get; set; }
         public ApagarCategoriaModel(ApplicationDbContext context)
         {
             _context = context;
@@ -19,6 +20,7 @@ namespace HabitAqui.Areas.Identity.Pages.Account.Manage
         public void OnGet(int id)
         {
             CategoriaId = id;
+            CategoriaNome = _context.Categorias.Find(id).Nome;
         }
 
         public async Task<IActionResult> OnPostAsync()
@@ -27,6 +29,15 @@ namespace HabitAqui.Areas.Identity.Pages.Account.Manage
 
             if (categoria != null)
             {
+                var isCategoriaUsed = await _context.Habitacoes
+                    .AnyAsync(h => h.Categorias != null && h.Categorias.Any(c => c.CategoriaId == CategoriaId));
+
+                if (isCategoriaUsed)
+                {
+                    ModelState.AddModelError("", "Não é possível apagar a categoria porque está associada a uma ou mais habitações.");
+                    return Page(); 
+                }
+
                 _context.Categorias.Remove(categoria);
                 await _context.SaveChangesAsync();
             }
@@ -35,3 +46,4 @@ namespace HabitAqui.Areas.Identity.Pages.Account.Manage
         }
     }
 }
+
